+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 
 ### Changes
 
-* Use Android plugin for Gradle v.7.3.1.
-* Use Kotlin v.1.7.21.
+* Use targetSdkVersion 33, compileSdkVersion 33.
+* Use Android plugin for Gradle v.7.4.2.
+* Use Kotlin v.1.8.10.
 * Use constraintlayout v.2.1.4.